maxCounters 100% 100%
algo hint from comments
So max counter operation is done after of if needed.  But bottom line it is outside the main for loop. So it's O(N+M).
We track the min and update a counter if it's below the min and then after anything that never got upped gets brought up to the min.